Revert "Refaktor til competenceReview og flere metoder"
This reverts commit 2daccd6b9638b50667d40f9c428e5f66204ba444.
