+++ SET_gruppe_13/src/main/java/no/hiof/g13/models/Question.java
@@ -3,13 +3,13 @@ package no.hiof.g13.models;
 import java.util.HashMap;
 
 public class Question {
-    private String questionType;
+   /* private int questionId;
     private String questionDescription;
     private HashMap<String, Integer> textAnswers;
     private HashMap<Integer, Integer> numAnswers;
 
-    public String getQuestionType() {
-        return this.questionType;
+    public int getQuestionId() {
+        return this.questionId;
     }
 
     public String getQuestionDescription() {
@@ -23,5 +23,5 @@ public class Question {
     }
     public HashMap<Integer, Integer> getNumAnswers() {
         return this.numAnswers;
-    }
+    } */
 }
